Changes were made to improve efficency
Trying to reduce memory usage and hopefully be a little more future
proof by loading in all of the items into a db, and only holding items
in a dict rather than all. This pretty much removes the need for
eve_items at all so any function that was passed or take that has been
changed.

- Fixed type in README.md
- Moved eve_items.csv to types .json
- Fixed code to only load input items rather than all into a dict