Adding History Averages
So, this is more of a part one. It kinda works now, but we're going to
polish it with the next one. the idea is to pretty much get average
information of items that are requested like average sold per day and
average prices. These things are pulled from eve crest

- Created way to get average sold
- added more information to sheets