Organization improvements
In order to move into the next phase of what I'm kinda hoping to see
this program go to, I think it's important that I tried to organize my
code a little bit and break it up. I hope this orginization makes it
easier for me to logically think through the program and make
improvements as needed.

- created crest file for crest calls
- created evecentral file for evecentral calls
- added data reports and templates file for further organization. In the
future I may add more templates and things like that so we'll see how
that plays out
- moved crest calls and eve_central calls to the crest and evecentral
files as well as removed deprecated code like make_urls since it was
moved into crest and evecentral files