Trying to make crest calls all threaded
So, the original implmentation of the get_Average_data had both the
class and defintion in it so I took out the threading part and broke it
into a new function that can now be used by any requests to it in order
to make things a little easier in the future for crest calls. I haven't
tested the new implmentation, but doing all of this threading has over
halved the time it takes this report to run. Hopefully in the future as
I mess with more threading I can hopefully make better code and or get a
better solution to bring it down any more.

- Created way for all requests to be threaded
- modified paths to data files in eve_trader to line up with changes of
  moving to data files