Back to Doing basic price info
First real update after. It's now at making price requests. With a
little more work I should be able to get multiple regions started, and a
little work after that I should be able to get shipping to work again.
I'm removing files that don't pertain to this and going to push this to
the main branch since I'm basically scrapping the previous version of
this.